+++ src/arbol/expresiones/aritmeticas/Multiplicacion.java
@@ -44,7 +44,7 @@ public class Multiplicacion extends Expresion {
             case entero :
                 switch (exp2.tipo.tipo) {
                     case entero :
-                        int porint = Integer.parseInt(str1) * Integer.parseInt(str2);
+                        int porint = (int)Double.parseDouble(str1) * (int)Double.parseDouble(str2);
                         literal= new Literal(new Tipo (Tipo.EnumTipo.entero) , porint);
                         break;
                     case doble :
